refactoring des tests
