+++ tests/Security/RegistrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Tests\Controller;
+namespace App\Tests\Security;
 
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\HttpFoundation\Request;
@@ -8,33 +8,8 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use App\Entity\User;
 
-class UserControllerTest extends WebTestCase
+class RegistrationTest extends WebTestCase
 {
-    /**
-     * @test
-     */
-    public function user_should_be_displayed()
-    {
-        $client = static::createClient();
-
-        /** @var EntityManagerInterface $entityManager */
-        $entityManager = $client->getContainer()->get("doctrine.orm.entity_manager");
-
-        $user = $entityManager->getRepository(User::class)->findOneBy([]);
-        $client->loginUser($user);
-
-        /** @var UrlGeneratorInterface $urlGenerator */
-        $urlGenerator = $client->getContainer()->get("router");
-
-        $crawler = $client->request(
-            Request::METHOD_GET,
-            $urlGenerator->generate("user_list")
-        );
-
-        $this->assertResponseStatusCodeSame(Response::HTTP_OK);
-        $this->assertCount(14, $crawler->filter('.user-infos'));
-    }
-
     /**
      * @test
      */
@@ -299,45 +274,5 @@ class UserControllerTest extends WebTestCase
         $client->submit($form);
 
         $this->assertSelectorTextContains('html', 'L\'email est déjà utilisé.');
-    }
-
-    /**
-     * @test
-     */
-    public function existant_user_should_be_edited()
-    {
-        $client = static::createClient();
-
-        /** @var UrlGeneratorInterface $urlGenerator */
-        $urlGenerator = $client->getContainer()->get("router");
-
-        /** @var EntityManagerInterface $entityManager */
-        $entityManager = $client->getContainer()->get("doctrine.orm.entity_manager");
-
-        $user = $entityManager->getRepository(User::class)->find(1);
-        $client->loginUser($user);
-
-        $crawler = $client->request(
-            Request::METHOD_GET,
-            $urlGenerator->generate("user_edit", ["id" => $user->getId(1)])
-        );
-
-        $form = $crawler->filter('form[name=user]')->form([
-            'user[username]' => 'new edited user',
-            'user[plainPassword][first]' => 'passworD1!',
-            'user[plainPassword][second]' => 'passworD1!',
-            'user[email]' => 'newuser@sf.com',
-        ]);
-
-        $client->submit($form);
-
-        $editedUser = $entityManager->getRepository(User::class)->find(1);
-
-        //comparer le changement d'état
-        $this->assertNotSame($user, $editedUser);
-
-        $this->assertResponseStatusCodeSame(Response::HTTP_FOUND);
-        $client->followRedirect();
-        $this->assertSelectorTextContains('td', 'new edited user');
     }
 }
