+++ tests/User/ShowTest.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace App\Tests\User;
+
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use App\Entity\User;
+
+class ShowTest extends WebTestCase
+{
+    /**
+     * @test
+     */
+    public function user_should_be_displayed()
+    {
+        $client = static::createClient();
+
+        /** @var EntityManagerInterface $entityManager */
+        $entityManager = $client->getContainer()->get("doctrine.orm.entity_manager");
+
+        $user = $entityManager->getRepository(User::class)->findOneBy([]);
+        $client->loginUser($user);
+
+        /** @var UrlGeneratorInterface $urlGenerator */
+        $urlGenerator = $client->getContainer()->get("router");
+
+        $crawler = $client->request(
+            Request::METHOD_GET,
+            $urlGenerator->generate("user_list")
+        );
+
+        $this->assertResponseStatusCodeSame(Response::HTTP_OK);
+        $this->assertCount(14, $crawler->filter('.user-infos'));
+    }
+}
